+++ test/unit/util_spec.js
@@ -17,7 +17,6 @@ import {
   bytesToString,
   createPromiseCapability,
   createValidAbsoluteUrl,
-  escapeString,
   getModificationDate,
   isArrayBuffer,
   string32,
@@ -241,14 +240,6 @@ describe("util", function () {
         expect(reason instanceof Error).toEqual(true);
         expect(reason.message).toEqual("reason");
       }
-    });
-  });
-
-  describe("escapeString", function () {
-    it("should escape (, ), \\n, \\r, and \\", function () {
-      expect(escapeString("((a\\a))\n(b(b\\b)\rb)")).toEqual(
-        "\\(\\(a\\\\a\\)\\)\\n\\(b\\(b\\\\b\\)\\rb\\)"
-      );
     });
   });
